agents/agent-archicteture/agent-architecture-agno/pyproject.toml
[project]
name = "agent-architecture-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno==2.0.0",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
]
